+++ romancal/photom/tests/test_photom.py
@@ -240,10 +240,12 @@ def test_photom_step_interface_spectroscopic(instrument, exptype):
     # Set photometric values for spectroscopic data
     wfi_image.meta.photometry.pixelarea_steradians = 2.31307642258977E-14 * u.steradian
     wfi_image.meta.photometry.pixelarea_arcsecsq = 0.000984102303070964 * u.arcsecond * u.arcsecond
-    wfi_image.meta.photometry.conversion_megajanskys = None
-    wfi_image.meta.photometry.conversion_megajanskys_uncertainty = None
-    wfi_image.meta.photometry.conversion_microjanskys = None
-    wfi_image.meta.photometry.conversion_microjanskys_uncertainty = None
+    wfi_image.meta.photometry.conversion_megajanskys = -99999 * u.megajansky / u.steradian
+    wfi_image.meta.photometry.conversion_megajanskys_uncertainty = -99999 * u.megajansky / \
+                                                                   u.steradian
+    wfi_image.meta.photometry.conversion_microjanskys = -99999 * u.microjansky / u.arcsecond ** 2
+    wfi_image.meta.photometry.conversion_microjanskys_uncertainty = -99999 * u.microjansky / \
+                                                                    u.arcsecond ** 2
 
     # Create input model
     wfi_image_model = ImageModel(wfi_image)
@@ -258,20 +260,10 @@ def test_photom_step_interface_spectroscopic(instrument, exptype):
     # Test that the data has not changed
     assert (np.allclose(result.data, wfi_image_model.data, rtol=1.e-7))
 
-    # Test that keywords are properly preserved
+    # Test that keywords are properly overwritten
     assert result.meta.photometry.conversion_megajanskys is None
     assert result.meta.photometry.conversion_microjanskys is None
     assert result.meta.photometry.conversion_megajanskys_uncertainty is None
     assert result.meta.photometry.conversion_microjanskys_uncertainty is None
-
-    # Set reference pixel areas
-    area_ster = 2.31307642258977E-14 * u.steradian
-    area_a2 = 0.000984102303070964 * u.arcsecond * u.arcsecond
-
-    # Tests for pixel areas
-    assert (np.isclose(result.meta.photometry.pixelarea_steradians.value,
-                        area_ster.value, atol=1.e-7))
-    assert result.meta.photometry.pixelarea_steradians.unit == area_ster.unit
-    assert (np.isclose(result.meta.photometry.pixelarea_arcsecsq.value,
-                        area_a2.value, atol=1.e-7))
-    assert result.meta.photometry.pixelarea_arcsecsq.unit == area_a2.unit
+    assert result.meta.photometry.pixelarea_steradians is None
+    assert result.meta.photometry.pixelarea_arcsecsq is None